+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Separate phantom argument from the definitions of `fset`, `fmap` and `fperm`
 
-- Add `ordType` instance for `mathcomp.choice.GenTree.tree`.
+- Add `ordType` instances for `mathcomp.choice.GenTree.tree` and `tuple`.
 
 - Add more implicit arguments for `fsubsetP`, `fset2P`, `imfsetP`, `dommP`,
   `dommPn`, `codommP` and `codommPn`.